Move tests from integration to unit spec
All of the logic for how to calculate a users favorite programming
language is going to be handled inside fetch_user obj. That's why
the tests were moved from the integration spec to the unit spec for
fetch_user.

The integration spec should check that a user can treat our
application like a command line application.